Handle lookup of ILogger member respecting backwards compatibility with base logging generator
