+++ src/Generators/Microsoft.Gen.Logging/Parsing/Parser.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Runtime.InteropServices.ComTypes;
 using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -32,7 +33,7 @@ internal sealed partial class Parser
     /// Gets the set of logging types containing methods to output.
     /// </summary>
     [SuppressMessage("Maintainability", "CA1505:Avoid unmaintainable code", Justification = "Fix this in a follow-up")]
-    public IReadOnlyList<LoggingType> GetLogTypes(IEnumerable<TypeDeclarationSyntax> types)
+    public IReadOnlyList<LoggingType> GetLogTypes(IEnumerable<TypeDeclarationSyntax> classes)
     {
         Action<DiagnosticDescriptor, Location?, object?[]?> diagReport = Diag; // Keeping one instance of the delegate
         var symbols = SymbolLoader.LoadSymbols(_compilation, diagReport);
@@ -48,9 +49,11 @@ internal sealed partial class Parser
         var parameterSymbols = new Dictionary<LoggingMethodParameter, IParameterSymbol>();
 
         // we enumerate by syntax tree, to minimize the need to instantiate semantic models (since they're expensive)
-        foreach (var group in types.GroupBy(x => x.SyntaxTree))
+        foreach (var group in classes.GroupBy(x => x.SyntaxTree))
         {
-            SemanticModel? sm = null;
+            SyntaxTree syntaxTree = group.Key;
+            SemanticModel? sm = _compilation.GetSemanticModel(syntaxTree);
+
             foreach (var typeDec in group)
             {
                 // stop if we're asked to
@@ -63,9 +66,15 @@ internal sealed partial class Parser
                 ISymbol? secondLoggerMember = null;
 
                 ids.Clear();
-                foreach (var method in typeDec.Members.Where(m => m.IsKind(SyntaxKind.MethodDeclaration)).Cast<MethodDeclarationSyntax>())
+
+                foreach (MemberDeclarationSyntax member in typeDec.Members)
                 {
-                    sm ??= _compilation.GetSemanticModel(typeDec.SyntaxTree);
+                    var method = member as MethodDeclarationSyntax;
+                    if (method == null)
+                    {
+                        // we only care about methods
+                        continue;
+                    }
 
                     var attrLoc = GetLoggerMessageAttribute(method, sm, symbols);
                     if (attrLoc == null)
@@ -200,44 +209,36 @@ internal sealed partial class Parser
 
                     if (keepMethod)
                     {
-                        if (lm.IsStatic)
+                        if (lm.IsStatic && !parsingState.FoundLogger)
                         {
-                            if (!parsingState.FoundLogger)
+                            Diag(DiagDescriptors.MissingLoggerParameter, method.ParameterList.GetLocation(), lm.Name);
+                            keepMethod = false;
+                        }
+                        else if (!lm.IsStatic && parsingState.FoundLogger)
+                        {
+                            Diag(DiagDescriptors.LoggingMethodShouldBeStatic, method.Identifier.GetLocation());
+                        }
+                        else if (!lm.IsStatic && !parsingState.FoundLogger)
+                        {
+                            if (loggerMember == null)
                             {
-                                Diag(DiagDescriptors.MissingLoggerParameter, method.ParameterList.GetLocation(), lm.Name);
+                                (loggerMember, secondLoggerMember, loggerMemberNullable) = FindLoggerMember(sm, typeDec, symbols.ILoggerSymbol);
+                            }
+
+                            if (secondLoggerMember != null)
+                            {
+                                Diag(DiagDescriptors.MultipleLoggerMembers, secondLoggerMember.GetLocation(), typeDec.Identifier.Text);
                                 keepMethod = false;
                             }
-                        }
-                        else
-                        {
-                            if (!parsingState.FoundLogger)
+                            else if (loggerMember == null)
                             {
-                                if (loggerMember == null)
-                                {
-                                    (loggerMember, secondLoggerMember, loggerMemberNullable) = FindMember(sm, typeDec, symbols.ILoggerSymbol);
-                                }
-
-                                if (secondLoggerMember != null)
-                                {
-                                    Diag(DiagDescriptors.MultipleLoggerMembers, secondLoggerMember.GetLocation(), typeDec.Identifier.Text);
-                                    keepMethod = false;
-                                }
-                                else if (loggerMember == null)
-                                {
-                                    Diag(DiagDescriptors.MissingLoggerMember, method.Identifier.GetLocation(), typeDec.Identifier.Text);
-                                    keepMethod = false;
-                                }
-                                else
-                                {
-                                    lm.LoggerMember = loggerMember;
-                                    lm.LoggerMemberNullable = loggerMemberNullable;
-                                }
+                                Diag(DiagDescriptors.MissingLoggerMember, method.Identifier.GetLocation(), typeDec.Identifier.Text);
+                                keepMethod = false;
                             }
-
-                            // Show this warning only if other checks passed
-                            if (keepMethod && parsingState.FoundLogger)
+                            else
                             {
-                                Diag(DiagDescriptors.LoggingMethodShouldBeStatic, method.Identifier.GetLocation());
+                                lm.LoggerMember = loggerMember;
+                                lm.LoggerMemberNullable = loggerMemberNullable;
                             }
                         }
 
@@ -617,67 +618,125 @@ internal sealed partial class Parser
         return null;
     }
 
-    private (string? member, ISymbol? secondMember, bool isNullable) FindMember(SemanticModel sm, TypeDeclarationSyntax classDec, ITypeSymbol symbol)
+    private (string? member, ISymbol? secondMember, bool isNullable) FindLoggerMember(SemanticModel sm, TypeDeclarationSyntax classDec, ITypeSymbol loggerSymbol)
     {
-        string? member = null;
+        string? loggerMember = null;
         bool isNullable = false;
 
-        INamedTypeSymbol? type = sm.GetDeclaredSymbol(classDec);
-        var originType = type!;
-        while (type != null)
-        {
-            foreach (var m in type.GetMembers())
-            {
-                if (m is IPropertySymbol p)
-                {
-                    var gm = p.GetMethod;
-                    if (gm != null)
-                    {
-                        if (ParserUtilities.IsBaseOrIdentity(gm.ReturnType, symbol, _compilation))
-                        {
-                            if (!sm.Compilation.IsSymbolAccessibleWithin(gm, originType, type))
-                            {
-                                continue;
-                            }
+        INamedTypeSymbol? classType = sm.GetDeclaredSymbol(classDec, _cancellationToken);
+        INamedTypeSymbol? currentClassType = classType;
+        bool onMostDerivedType = true;
 
-                            if (member == null)
-                            {
-                                member = p.Name;
-                                isNullable = gm.ReturnType.NullableAnnotation == NullableAnnotation.Annotated;
-                            }
-                            else
-                            {
-                                return (null, p, isNullable);
-                            }
-                        }
+#pragma warning disable S125 // Sections of code should not be commented out
+        // We keep track of the names of all non-logger fields, since they prevent referring to logger
+        // primary constructor parameters with the same name. Example:
+        // partial class C(ILogger logger)
+        // {
+        //     private readonly object logger = logger;
+        //
+        //     [LoggerMessage(EventId = 0, Level = LogLevel.Debug, Message = ""M1"")]
+        //     public partial void M1(); // The ILogger primary constructor parameter cannot be used here.
+        // }
+#pragma warning restore S125 // Sections of code should not be commented out
+
+        HashSet<string> shadowedNames = new(StringComparer.Ordinal);
+
+        while (currentClassType is { SpecialType: not SpecialType.System_Object })
+        {
+            foreach (ISymbol ms in currentClassType.GetMembers())
+            {
+                // we support both fields and properties
+                ITypeSymbol? typeSymbol = ms switch
+                {
+                    IFieldSymbol fs => fs.Type,
+                    IPropertySymbol ps => ps.Type,
+                    _ => default,
+                };
+
+                if (typeSymbol is null)
+                {
+                    continue;
+                }
+
+                if (!onMostDerivedType && ms.DeclaredAccessibility == Accessibility.Private)
+                {
+                    continue;
+                }
+
+                if (!ms.CanBeReferencedByName)
+                {
+                    continue;
+                }
+
+                if (ParserUtilities.IsBaseOrIdentity(typeSymbol, loggerSymbol, _compilation))
+                {
+                    if (loggerMember == null)
+                    {
+                        loggerMember = ms.Name;
+                        isNullable = typeSymbol.NullableAnnotation == NullableAnnotation.Annotated;
+                    }
+                    else if (!onMostDerivedType)
+                    {
+                        // we found a public logger member on a base class, prefer the existing one
+                        // because it's more likely to be the one the user intended
+                        // the dotnet logging generator doesn't support ILogger properties so this matches that behavior
+                        continue;
+                    }
+                    else
+                    {
+                        return (null, ms, isNullable);
                     }
                 }
-                else if (m is IFieldSymbol f)
+                else
                 {
-                    if (f.AssociatedSymbol == null && ParserUtilities.IsBaseOrIdentity(f.Type, symbol, _compilation))
-                    {
-                        if (!sm.Compilation.IsSymbolAccessibleWithin(f, originType, type))
-                        {
-                            continue;
-                        }
-
-                        if (member == null)
-                        {
-                            member = f.Name;
-                            isNullable = f.Type.NullableAnnotation == NullableAnnotation.Annotated;
-                        }
-                        else
-                        {
-                            return (null, f, isNullable);
-                        }
-                    }
+                    _ = shadowedNames.Add(ms.Name);
                 }
             }
 
-            type = type.BaseType;
+            onMostDerivedType = false;
+            currentClassType = currentClassType.BaseType;
         }
 
-        return (member, null, isNullable);
+        // We prioritize fields over primary constructor parameters and avoid warnings if both exist.
+        if (loggerMember is not null)
+        {
+            return (loggerMember, null, isNullable);
+        }
+
+        IEnumerable<IMethodSymbol> primaryConstructors = classType!.InstanceConstructors
+            .Where(ic => ic.DeclaringSyntaxReferences
+                .Any(ds => ds.GetSyntax() is ClassDeclarationSyntax));
+
+        foreach (IMethodSymbol primaryConstructor in primaryConstructors)
+        {
+            foreach (IParameterSymbol parameter in primaryConstructor.Parameters)
+            {
+                if (ParserUtilities.IsBaseOrIdentity(parameter.Type, loggerSymbol, _compilation))
+                {
+                    if (shadowedNames.Contains(parameter.Name))
+                    {
+                        // Accessible fields always shadow primary constructor parameters,
+                        // so we can't use the primary constructor parameter,
+                        // even if the field is not a valid logger.
+                        Diag(DiagDescriptors.PrimaryConstructorParameterLoggerHidden, parameter.GetLocation(), classDec.Identifier.Text);
+
+                        continue;
+                    }
+
+                    if (loggerMember == null)
+                    {
+                        loggerMember = parameter.Name;
+                        isNullable = parameter.Type.NullableAnnotation == NullableAnnotation.Annotated;
+                    }
+                    else
+                    {
+                        return (null, parameter.Type, isNullable);
+                    }
+                }
+            }
+        }
+
+        return (loggerMember, null, isNullable);
     }
 
     private void Diag(DiagnosticDescriptor desc, Location? location, params object?[]? messageArgs)
