+++ src/Shared/DiagnosticIds/DiagnosticIds.cs
@@ -87,6 +87,7 @@ internal static class DiagnosticIds
         internal const string LOGGEN035 = nameof(LOGGEN035);
         internal const string LOGGEN036 = nameof(LOGGEN036);
         internal const string LOGGEN037 = nameof(LOGGEN037);
+        internal const string LOGGEN038 = nameof(LOGGEN038);
     }
 
     internal static class Metrics
